+++ NEWS.md
@@ -1,4 +1,5 @@
 ## 1.3.0 ##
+* Set `.libPaths()` in subprocess to match those in calling process (#140, #147).
 * Move devtools dependency to suggests, only needed on windows
 * move htmltools to suggests